[fix]: Moved splash screen logic from MainActivity.kt to InitActivity.kt
