+++ app/src/main/res/values/colors.xml
@@ -2,11 +2,9 @@
 <resources>
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
-
     <color name="background">#F2FAF8</color>
-
     <color name="init_block_bg">#FFFFFF</color>
     <color name="init_block_border">#CAC4D0</color>
     <color name="init_button_bg">#556B8F</color>
-    <color name="background">#F2FAF8</color>
+
 </resources>
